+++ packages/firebase_auth/firebase_auth_dart/lib/src/api.dart
